Various code cleanup, emphasis on includes
* ObservableProperty and ObservableWritableProperty are split into
  separate file in libsoftwarecontainer.
* SignalConnectionsHandler is split out in separate file.
* softwarecontainer-common.h doesn't include everything anymore, which
  means that includes are placed where needed instead. This is only a
  partial cleanup, however.
