+++ libsoftwarecontainer/include/observableproperty.h
@@ -1,0 +1,90 @@
+/*
+ * Copyright (C) 2016 Pelagicore AB
+ *
+ * Permission to use, copy, modify, and/or distribute this software for
+ * any purpose with or without fee is hereby granted, provided that the
+ * above copyright notice and this permission notice appear in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL
+ * WARRANTIES WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR
+ * BE LIABLE FOR ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES
+ * OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS,
+ * WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION,
+ * ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS
+ * SOFTWARE.
+ *
+ * For further information see LICENSE
+ */
+
+#include <functional>
+#include <vector>
+
+template<typename Type>
+class ObservableProperty
+{
+public:
+    typedef std::function<void (const Type &)> Listener;
+
+    ObservableProperty(Type &value) :
+        m_value(value)
+    {
+    }
+
+    void addListener(Listener listener)
+    {
+        m_listeners.push_back(listener);
+    }
+
+    operator const Type &() {
+        return m_value;
+    }
+
+protected:
+    std::vector<Listener> m_listeners;
+
+private:
+    const Type &m_value;
+
+};
+
+template<typename Type>
+class ObservableWritableProperty :
+    public ObservableProperty<Type>
+{
+public:
+    ObservableWritableProperty(Type value) :
+        ObservableProperty<Type>(m_value), m_value(value)
+    {
+    }
+
+    ObservableWritableProperty() :
+        ObservableProperty<Type>(m_value)
+    {
+    }
+
+    void setValueNotify(Type value)
+    {
+        m_value = value;
+        for (auto &listener : ObservableProperty<Type>::m_listeners) {
+            listener(getValue());
+        }
+    }
+
+    const Type &getValue() const
+    {
+        return m_value;
+    }
+
+    ObservableWritableProperty &operator=(const Type &type)
+    {
+        m_value = type;
+        return *this;
+    }
+
+
+private:
+    Type m_value;
+
+};
+
